Moved keytransparency library client package to com.google.keytransparency.client and removed unnecessary statements from manifest.
