+++ keytransparency/src/main/AndroidManifest.xml
@@ -1,12 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-
-    package="com.google.keytransparency">
+    package="com.google.keytransparency.client">
 
     <uses-permission android:name="android.permission.INTERNET" />
-
-    <application android:allowBackup="true" android:label="@string/app_name"
-        android:supportsRtl="true">
-
-    </application>
 
 </manifest>
