+++ keytransparency/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">keytransparency</string>
+    <string name="app_name">TODO_REMOVE_THIS_FILE</string>
 </resources>
